requirements.txt
reflex>=0.4.0
openai>=1.12.0
pandas
streamlit
langchain
pypresence
google-generativeai
python-dotenv
streamlit
pillow